fix(ci): refactor to fix `sync-conda-lock`
