+++ python/utils.py
@@ -7,10 +7,6 @@ from pathlib import Path
 from shutil import rmtree
 from typing import Any, Callable, Final, Generator, Iterable, Iterator, Optional, Union
 
-from numpy import array, random
-from pandas import to_datetime
-from xarray import DataArray
-
 DateType = Union[date, str]
 DATE_FORMAT_STR: Final[str] = "%Y%m%d"
 ISO_DATE_FORMAT_STR: Final[str] = "%Y-%m-%d"
@@ -30,17 +26,6 @@ DEFAULT_CONDA_LOCK_KWARGS: Final[dict[str, str | float | bool]] = {
 }
 
 GITHUB_ACTIONS_ARCHITECTURE: Final[str] = "linux-64"
-
-
-GLASGOW_COORDS: Final[tuple[float, float]] = (55.86279, -4.25424)
-MANCHESTER_COORDS: Final[tuple[float, float]] = (53.48095, -2.23743)
-LONDON_COORDS: Final[tuple[float, float]] = (51.509865, -0.118092)
-CITY_COORDS: Final[dict[str, tuple[float, float]]] = {
-    "Glasgow": GLASGOW_COORDS,
-    "Manchester": MANCHESTER_COORDS,
-    "London": LONDON_COORDS,
-}
-"""Coordinates of Glasgow, Manchester and London as `(lon, lat)` `tuples`."""
 
 
 def ensure_date(date_to_check: DateType, format_str: str = DATE_FORMAT_STR) -> date:
@@ -324,83 +309,6 @@ def path_iterdir(
             raise error
         else:
             return
-
-
-def xarray_example(
-    start_date: DateType,
-    end_date: DateType,
-    coordinates: dict[str, tuple[float, float]] = CITY_COORDS,
-    skip_dates: Iterable[date] | None = None,
-    random_seed_int: int | None = None,
-    **kwargs,
-) -> DataArray:
-    """Generate spatial and temporal `xarray` objects.
-
-    Parameters
-    ----------
-    start_date
-        Start of time series.
-    end_date
-        End of time series (by default not inclusive).
-    coordinates
-        A `dict` of region name `str` to `tuple` of
-        `(lon, lat)` form.
-    skip_dates
-        A list of `date` objects to drop/skip between
-        `start_date` and `end_date`.
-    kwargs
-        Additional parameters to pass to `date_range_generator`.
-
-    Returns
-    -------
-    A `DataArray` of `start_date` to `end_date` date range a
-    random variable for coordinates regions
-    (Glasgow, Manchester and London as default).
-
-    Examples
-    --------
-    >>> xarray_example('1980-11-30', '1980-12-5')
-    <xarray.DataArray (time: 5, space: 3)> Size: 120B
-    array([[..., ..., ...],
-           [..., ..., ...],
-           [..., ..., ...],
-           [..., ..., ...],
-           [..., ..., ...]])
-    Coordinates:
-      * time     (time) datetime64[ns] 40B 1980-11-30 1980-12-01 ... 1980-12-04
-      * space    (space) <U10 120B 'Glasgow' 'Manchester' 'London'
-    """
-    dates: list[DateType] = list(
-        date_range_generator(
-            start_date=start_date,
-            end_date=end_date,
-            start_format_str=ISO_DATE_FORMAT_STR,
-            end_format_str=ISO_DATE_FORMAT_STR,
-            skip_dates=skip_dates,
-            **kwargs,
-        )
-    )
-    if isinstance(random_seed_int, int):
-        random.seed(random_seed_int)  # ensure results are predictable
-    data: array = random.rand(len(dates), len(coordinates))
-    spaces: list[str] = list(coordinates.keys())
-    # If useful, add lat/lon (currently not working)
-    # lat: list[float] = [coord[0] for coord in coordinates.values()]
-    # lon: list[float] = [coord[1] for coord in coordinates.values()]
-    return DataArray(
-        data,
-        coords=[
-            to_datetime(dates),
-            spaces,
-        ],
-        dims=[
-            "time",
-            "space",
-        ],
-        # If useful, add lat/lon (currently not working)
-        # coords=[dates, spaces, lon, lat],
-        # dims=["time", "space", "lon", "lat"]
-    )
 
 
 def run_conda_lock(
